projects/scripts/pyproject.toml
[project]
name = "uoft_scripts"
description = "a collection of scripts and one-off tools"
authors = [
  { name = "Alex Tremblay", email = "alex.tremblay@utoronto.ca" },
  { name = "University of Toronto", email = "alex.tremblay@utoronto.ca" },
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
dynamic = ["version"]
dependencies = [
  "uoft_core",
  "uoft_bluecat",
  "uoft_ssh",
  "uoft_librenms",
  "uoft_paloalto",
  "pynautobot >= 2.1.1",
  "jinja2 >= 3.0",
  "deepdiff >= 7.0",
  "python-box >= 7.2",
  "glom >= 23.5",
  "pydantic >= 1.9.0",
  "prompt-toolkit >= 3.0.19",
  "rich >= 10.7.0",
  "pyquery >= 1.4.0",
  "ldap3 >= 2.9",
  "nornir_napalm >= 0.4",
  "nornir_netmiko >= 1.0",
  "nornir_nautobot >= 3.2",
  "sqlmodel >= 0.0.22",
  "psycopg >= 3.2",
  "grpcio >= 1.71",
  "grpcio_reflection >= 1.71",
  "protobuf >= 5",
  "cloudvision >= 1.23"
]

[project.scripts]
uoft-scripts = "uoft_scripts.cli:cli"

[build-system]
requires = ["pdm-backend", "debugpy"]
build-backend = "pdm.backend"

[tool.pdm.version]
source = "scm"

[tool.pdm.build]
custom-hook = "pyproject.py"
includes = ["uoft_scripts"]
excludes = ["**/tests", "**/__pycache__"]

[tool.uv]
dev-dependencies = [
    "grpcio-tools>=1.71.0",
]
